refactor: simplify script history event handling by extracting listeners into a dedicated function
